common/badger: Move common BadgerDB integration helpers here
There's no reason why the log adapter needs to live deep in the MKVS
implementation, and there is no good reason why there needs to be 2
copies of the value log GC worker.

While I'm here, also add value log GC-ing to the runtime and node
persistent storage.
